Fast spacetime space implemented, hierarchy of spacetimes spaces modified.

+++ cpp/include/besthea/spacetime_cluster_tree.h
@@ -90,26 +90,40 @@ class besthea::mesh::spacetime_cluster_tree {
 
  private:
   using full_matrix_type = besthea::linear_algebra::full_matrix;
-  std::vector< full_matrix_type > _m2m_matrices_t_left;  //! left temporal 
-      //! m2m matrices stored levelwise
-  std::vector< full_matrix_type > _m2m_matrices_t_right; //! right temporal 
-      //! m2m matrices stored levelwise
-  lo _temp_order = 5; //! degree of interpolation polynomials in time for pFMM 
-  std::vector< vector_type > _m2m_coeffs_s_dim_0_left;  //! left spatial 
-      //! m2m matrices along dimension 0 stored levelwise
-  std::vector< vector_type > _m2m_coeffs_s_dim_0_right;  //! right spatial 
-      //! m2m matrices along dimension 0 stored levelwise
-  std::vector< vector_type > _m2m_coeffs_s_dim_1_left;  //! left spatial 
-      //! m2m matrices along dimension 1 stored levelwise
-  std::vector< vector_type > _m2m_coeffs_s_dim_1_right;  //! right spatial 
-      //! m2m matrices along dimension 1 stored levelwise
-  std::vector< vector_type > _m2m_coeffs_s_dim_2_left;  //! left spatial 
-      //! m2m matrices along dimension 2 stored levelwise
-  std::vector< vector_type > _m2m_coeffs_s_dim_2_right;  //! right spatial 
-      //! m2m matrices along dimension 2 stored levelwise
-  lo _spat_order = 5; //! degree of Chebyshev polynomials for expansion in
-      //! space in pFMM
-  
+  std::vector< full_matrix_type >
+    _m2m_matrices_t_left;  //! left temporal
+                           //! m2m matrices stored levelwise
+  std::vector< full_matrix_type >
+    _m2m_matrices_t_right;  //! right temporal
+                            //! m2m matrices stored levelwise
+  lo _temp_order = 5;  //! degree of interpolation polynomials in time for pFMM
+  std::vector< vector_type >
+    _m2m_coeffs_s_dim_0_left;  //! left spatial
+                               //! m2m matrices along dimension 0 stored
+                               //! levelwise
+  std::vector< vector_type >
+    _m2m_coeffs_s_dim_0_right;  //! right spatial
+                                //! m2m matrices along dimension 0 stored
+                                //! levelwise
+  std::vector< vector_type >
+    _m2m_coeffs_s_dim_1_left;  //! left spatial
+                               //! m2m matrices along dimension 1 stored
+                               //! levelwise
+  std::vector< vector_type >
+    _m2m_coeffs_s_dim_1_right;  //! right spatial
+                                //! m2m matrices along dimension 1 stored
+                                //! levelwise
+  std::vector< vector_type >
+    _m2m_coeffs_s_dim_2_left;  //! left spatial
+                               //! m2m matrices along dimension 2 stored
+                               //! levelwise
+  std::vector< vector_type >
+    _m2m_coeffs_s_dim_2_right;  //! right spatial
+                                //! m2m matrices along dimension 2 stored
+                                //! levelwise
+  lo _spat_order = 5;  //! degree of Chebyshev polynomials for expansion in
+                       //! space in pFMM
+
   // TODO relocate the above members to pFMM matrix
   // TODO substitute hardcoded _temp_order and _spatial_order with settable ones
 
@@ -118,13 +132,13 @@ class besthea::mesh::spacetime_cluster_tree {
   time_cluster_tree * _time_tree;               //!< temporal tree
   space_cluster_tree * _space_tree;             //!< spatial tree
   spacetime_cluster * _root;                    //!< root of the cluster tree
-  lo _start_spatial_level;  //!< auxiliary variable determining the appropriate
-                            //!< starting level in the space cluster tree
-  lo _start_temporal_level; //!< auxiliary variable to determine in which level
-                            //!< the spatial refinement starts 
-                            //!< (meaningful only if _start_spatial_level = 0)
-  sc _s_t_coeff;    //!< coefficient to determine the coupling of the spatial
-                    //!< and temporal levels
+  lo _start_spatial_level;   //!< auxiliary variable determining the appropriate
+                             //!< starting level in the space cluster tree
+  lo _start_temporal_level;  //!< auxiliary variable to determine in which level
+                             //!< the spatial refinement starts
+                             //!< (meaningful only if _start_spatial_level = 0)
+  sc _s_t_coeff;  //!< coefficient to determine the coupling of the spatial
+                  //!< and temporal levels
 
   /**
    * Auxiliary method to get all spatial clusters on a given level
@@ -141,29 +155,29 @@ class besthea::mesh::spacetime_cluster_tree {
    * Aux for printing
    */
   void print_internal( spacetime_cluster * root ) {
-    if (root->get_level( ) == -1) {
+    if ( root->get_level( ) == -1 ) {
       // print temporal m2m matrices
       lo n_levels = _time_tree->get_levels( );
       std::cout << "printing exemplarily temporal m2m matrices" << std::endl;
       for ( lo curr_lev = 2; curr_lev < n_levels; ++curr_lev ) {
         std::cout << "printing m2m matrix on level " << curr_lev << std::endl;
-        for ( lo j = 0; j <= _temp_order; ++ j ) {
-          for ( lo k = 0; k <= _temp_order; ++ k )
-            printf( "%.4f ", _m2m_matrices_t_right[curr_lev](j, k) );
+        for ( lo j = 0; j <= _temp_order; ++j ) {
+          for ( lo k = 0; k <= _temp_order; ++k )
+            printf( "%.4f ", _m2m_matrices_t_right[ curr_lev ]( j, k ) );
           std::cout << std::endl;
         }
       }
       // print spatial m2m coefficients
       n_levels = _space_tree->get_levels( );
       std::cout << "printing exemplarily spatial m2m coefficients" << std::endl;
-      for ( lo curr_lev = 0; curr_lev < n_levels; ++ curr_lev ) {
+      for ( lo curr_lev = 0; curr_lev < n_levels; ++curr_lev ) {
         std::cout << "printing m2m coeffs on level " << curr_lev << std::endl;
         std::cout << "curr_lev = " << curr_lev << std::endl;
-        for ( lo j = 0; j <= _spat_order; ++ j ) {
-          for ( lo k = 0; k <= _spat_order; ++ k )
-            printf( "(%d, %d): %.4f ", j, k, 
-              _m2m_coeffs_s_dim_0_left[ curr_lev ][ 
-               j * ( _spat_order + 1 ) + k ] );
+        for ( lo j = 0; j <= _spat_order; ++j ) {
+          for ( lo k = 0; k <= _spat_order; ++k )
+            printf( "(%ld, %ld): %.4f ", (long) j, (long) k,
+              _m2m_coeffs_s_dim_0_left[ curr_lev ]
+                                      [ j * ( _spat_order + 1 ) + k ] );
           std::cout << std::endl;
         }
       }
@@ -178,32 +192,29 @@ class besthea::mesh::spacetime_cluster_tree {
         print_internal( *it );
       }
   }
-  
+
   /*
    * Compute the temporal m2m matrices for all levels.
    */
   void set_temporal_m2m_matrices( );
-  
+
   /*
    * Compute the spatial m2m coefficients for all levels.
    */
   void set_spatial_m2m_coeffs( );
-  
+
   /*
    * Apply the temporal m2m operation for given parent and child moments
    */
-  void apply_temporal_m2m( full_matrix_type const & child_moment, 
-                           const lo level,
-                           const bool is_left_child, 
-                           full_matrix_type & parent_moment);
-  
+  void apply_temporal_m2m( full_matrix_type const & child_moment,
+    const lo level, const bool is_left_child,
+    full_matrix_type & parent_moment );
+
   /*
    * Apply the spatial m2m operation for given parent and child moments
    */
-  void apply_spatial_m2m( full_matrix_type const & child_moment, 
-                          const lo level,
-                          const slou octant, 
-                          full_matrix_type & parent_moment);
+  void apply_spatial_m2m( full_matrix_type const & child_moment, const lo level,
+    const slou octant, full_matrix_type & parent_moment );
 };
 
 #endif /* INCLUDE_BESTHEA_SPACETIME_CLUSTER_TREE_H_ */
